requirements.txt
gTTS==2.5.3
openai_whisper==20240930
protobuf==5.29.3
SpeechRecognition==3.10.4
streamlit==1.41.1
